Option #2: include asyncTransformer in package (passed thru config)
